started major rewrite to move absolutely everything around data chunks into the DataChunk API - will probably have to disentangle this into multiple smaller PRs
