+++ vector_cast.go
@@ -1,0 +1,167 @@
+package duckdb
+
+/*
+#include <stdlib.h>
+#include <duckdb.h>
+*/
+import "C"
+
+import (
+	"reflect"
+	"strconv"
+	"time"
+)
+
+func (vec *vector) tryCast(val any) (any, error) {
+	if val == nil {
+		return val, nil
+	}
+
+	// TODO: add all remaining types.
+
+	switch vec.duckdbType {
+	case C.DUCKDB_TYPE_UTINYINT:
+		return tryNumericCast[uint8](val, reflect.Uint8.String())
+	case C.DUCKDB_TYPE_TINYINT:
+		return tryNumericCast[int8](val, reflect.Int8.String())
+	case C.DUCKDB_TYPE_USMALLINT:
+		return tryNumericCast[uint16](val, reflect.Uint16.String())
+	case C.DUCKDB_TYPE_SMALLINT:
+		return tryNumericCast[int16](val, reflect.Int16.String())
+	case C.DUCKDB_TYPE_UINTEGER:
+		return tryNumericCast[uint32](val, reflect.Uint32.String())
+	case C.DUCKDB_TYPE_INTEGER:
+		return tryNumericCast[int32](val, reflect.Int32.String())
+	case C.DUCKDB_TYPE_UBIGINT:
+		return tryNumericCast[uint64](val, reflect.Uint64.String())
+	case C.DUCKDB_TYPE_BIGINT:
+		return tryNumericCast[int64](val, reflect.Int64.String())
+	case C.DUCKDB_TYPE_FLOAT:
+		return tryNumericCast[float32](val, reflect.Float32.String())
+	case C.DUCKDB_TYPE_DOUBLE:
+		return tryNumericCast[float64](val, reflect.Float64.String())
+	case C.DUCKDB_TYPE_BOOLEAN:
+		return tryPrimitiveCast[bool](val, reflect.Bool.String())
+	case C.DUCKDB_TYPE_VARCHAR:
+		return tryPrimitiveCast[string](val, reflect.String.String())
+	case C.DUCKDB_TYPE_BLOB:
+		return tryPrimitiveCast[[]byte](val, reflect.TypeOf([]byte{}).String())
+	case C.DUCKDB_TYPE_TIMESTAMP, C.DUCKDB_TYPE_TIMESTAMP_S, C.DUCKDB_TYPE_TIMESTAMP_MS,
+		C.DUCKDB_TYPE_TIMESTAMP_NS, C.DUCKDB_TYPE_TIMESTAMP_TZ, C.DUCKDB_TYPE_DATE:
+		return tryPrimitiveCast[time.Time](val, reflect.TypeOf(time.Time{}).String())
+	case C.DUCKDB_TYPE_UUID:
+		return tryPrimitiveCast[UUID](val, reflect.TypeOf(UUID{}).String())
+	case C.DUCKDB_TYPE_LIST:
+		return vec.tryCastList(val)
+	case C.DUCKDB_TYPE_STRUCT:
+		return vec.tryCastStruct(val)
+	}
+
+	return nil, getError(errDriver, nil)
+}
+
+func (*vector) canNil(val reflect.Value) bool {
+	switch val.Kind() {
+	case reflect.Chan, reflect.Func, reflect.Map, reflect.Pointer,
+		reflect.UnsafePointer, reflect.Interface, reflect.Slice:
+		return true
+	}
+	return false
+}
+
+func tryPrimitiveCast[T any](val any, expected string) (any, error) {
+	if v, ok := val.(T); ok {
+		return v, nil
+	}
+
+	goType := reflect.TypeOf(val)
+	return nil, castError(goType.String(), expected)
+}
+
+func tryNumericCast[T numericType](val any, expected string) (any, error) {
+	if v, ok := val.(T); ok {
+		return v, nil
+	}
+
+	// JSON unmarshalling uses float64 for numbers.
+	// We might want to add more implicit casts here.
+	switch v := val.(type) {
+	case float64:
+		return convertNumericType[float64, T](v), nil
+	}
+
+	goType := reflect.TypeOf(val)
+	return nil, castError(goType.String(), expected)
+}
+
+func (vec *vector) tryCastList(val any) ([]any, error) {
+	goType := reflect.TypeOf(val)
+	if goType.Kind() != reflect.Slice {
+		return nil, castError(goType.String(), reflect.Slice.String())
+	}
+
+	v := reflect.ValueOf(val)
+	list := make([]any, v.Len())
+	childVector := vec.childVectors[0]
+
+	for i := 0; i < v.Len(); i++ {
+		idx := v.Index(i)
+		if vec.canNil(idx) && idx.IsNil() {
+			list[i] = nil
+			continue
+		}
+
+		var err error
+		list[i], err = childVector.tryCast(idx.Interface())
+		if err != nil {
+			return nil, err
+		}
+	}
+	return list, nil
+}
+
+func (vec *vector) tryCastStruct(val any) (map[string]any, error) {
+	m, isMap := val.(map[string]any)
+
+	// Transform the struct into map[string]any.
+	if !isMap {
+		// Catch mismatching types.
+		goType := reflect.TypeOf(val)
+		if reflect.TypeOf(val).Kind() != reflect.Struct {
+			return nil, castError(goType.String(), reflect.Struct.String())
+		}
+
+		m = make(map[string]any)
+		v := reflect.ValueOf(val)
+		structType := v.Type()
+
+		for i := 0; i < structType.NumField(); i++ {
+			fieldName := structType.Field(i).Name
+			m[fieldName] = v.Field(i).Interface()
+		}
+	}
+
+	// Catch mismatching field count.
+	if len(m) != len(vec.childNames) {
+		return nil, structFieldError(strconv.Itoa(len(m)), strconv.Itoa(len(vec.childNames)))
+	}
+
+	// Cast child entries and return the map.
+	for i := 0; i < len(vec.childVectors); i++ {
+		childVector := vec.childVectors[i]
+		childName := vec.childNames[i]
+		v, ok := m[childName]
+
+		// Catch mismatching field names.
+		if !ok {
+			return nil, structFieldError("missing field", childName)
+		}
+
+		var err error
+		m[childName], err = childVector.tryCast(v)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return m, nil
+}
